+++ packages/ui/src/components/Float/index.ts
@@ -1,0 +1,2 @@
+export { default as Float } from './float';
+export * from './float.types';