+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @bako/ui
 
+## 0.3.2
+
+### Patch Changes
+
+- [`a3e01e0`](https://github.com/infinitybase/bako-ui/commit/a3e01e0201727956ed4dcc7ee8a9396818a893cf) Thanks [@Vitosoaresp](https://github.com/Vitosoaresp)! - fix radio component to show indicator
+
+## 0.3.1
+
+### Patch Changes
+
+- [`1500093`](https://github.com/infinitybase/bako-ui/commit/150009323542ef0c87855bb5f4a2984525028731) Thanks [@Vitosoaresp](https://github.com/Vitosoaresp)! - update badge recipe
+
+## 0.3.0
+
+### Minor Changes
+
+- [#13](https://github.com/infinitybase/bako-ui/pull/13) [`d87596a`](https://github.com/infinitybase/bako-ui/commit/d87596ab6ea9db1aa2009d76668b667f5809c1fa) Thanks [@Vitosoaresp](https://github.com/Vitosoaresp)! - add float component
+
+## 0.2.5
+
+### Patch Changes
+
+- [#11](https://github.com/infinitybase/bako-ui/pull/11) [`1176466`](https://github.com/infinitybase/bako-ui/commit/117646679e5e9e45f161694b8cec46cd9a49fcc9) Thanks [@Vitosoaresp](https://github.com/Vitosoaresp)! - improve avatar loading with skeleton and export select hooks
+
 ## 0.2.4
 
 ### Patch Changes